Fix reset of exception ports for which a server was not previously registered.
This fixes the case where we register for exception types (via
FATAL_EXCEPTION_MASK) that were not registered to an existing server.
The ports in question would be left registered to our exception server
after it was stopped, which would result in a hang in the case where an
exception occured.

The fix is to reset all ports; in the case where a server was previously
registered, we re-insert it. In the case where none was registered, we
set the exception handler server port to MACH_PORT_NULL.

It remains the case that deregistering a mach exception server should
not generally be done, for the reasons described in the
deregisterHandlerAndReturnError: method documentation.