Checkpoint commit: Begin unifying frame walker code across architectures, and drop PPC support.
This work is required to support merging in the DWARF, compact unwind,
stack scanning frame walkers, and to allow the implementation of
prologue scanning. PPC hasn't been supported in 2-3 years; due to
significant changes made here, is no longer reasonable to keep the
PPC sources in-tree.

This also has the side benefit of simplifying the Mach exception work,
as it creates a new unified thread state structure that is distinct from
the signal-specific ucontext_t.

This code is complete and functional, with one exception (which is why
this commit is marked with 'Checkpoint'):

In -[PLCrashLogWriterTests testWriteReport], the
plcrash_log_writer_write() call has been short-circuited to not supply
the expected ucontext_t. This will have to be resolved by removing the
minimal remaining dependencies on ucontext_t within the PLCrashLogWriter
APIs.

Checkpoint TODOs in this commit (there is only the one described above)
are marked with the THREAD_STATE_TODO flag.