Add a hackish workaround to simplify executing DemoCrash on a device.
On iOS 6.x, when using Xcode 4, returning *immediately* from main()
while a debugger is attached will cause an immediate launchd respawn of the
application without the debugger enabled.

This is not documented anywhere, and certainly occurs entirely by accident.
That said, it's enormously useful when performing integration tests on signal/exception
handlers, as it means we can use the standard Xcode build+run functionality without having
the debugger catch our signals (thus requiring that we manually relaunch the app after it has
installed).

This may break at any point in the future, in which case we can remove it and go back
to the old, annoying, and slow approach of manually relaunching the application. Or,
perhaps Apple will bless us with the ability to run applications without the debugger
enabled.