Implement Mach-O binary loading.
This includes a basic fat binary loader, which will load the binary that
matches the test process itself. The result is that loading will fail
should the test process be build for a target that the target binaries
are not.

Since each new architecture and platform requires careful consideration
prior to inclusion, failing fast is probably the best approach here.