Add support for returning mobject mappings shorter than the requested length.
Due to bugs in the update_dyld_shared_cache(1), the segment vmsize
defined in the Mach-O load commands may be invalid, and the
declared size may be unmappable.

This bug appears to be caused by a bug in computing the correct vmsize
when update_dyld_shared_cache(1) generates the single shared LINKEDIT
segment, and has been reported to Apple as rdar://13707406.