Modify the _plcrash_log_writer_write_curthread() assembly trampoline to support calling arbitrary callbacks, rather than only issueing a crash report write.
This allows us to use the function for other purposes, such as fetching
the current thread's state, without having to also write a full crash
report.

For the sake of minimizing changes in this commit, the function has not
yet been renamed or moved to suit its more generic usage.