Implement basic GNU eh_frame pointer encoding support.
This is based on the DWARF extensions documented in the Linux Standard
Base Core Specification 4.1, section 10.5, DWARF Extensions.

Since this document does not fully define the appropriate interpretation
of the encoded pointer values, actual implementations where consulted,
and the source for any informationed gleaned has been included in code
comments.

The following projects are (and will continue to be) consulted to clarify
ambiguious standard issues:
	gdb 7.6
	libunwind 1.1
	lldb-trunk (svn r183073)
	clang (git 69a8552f85c1b926320418cb98fe02988c79578f)
	Apple libunwind-35.1
	Apple ld64-134.9

Responsibility for any misinterpration of their implementation
rests entirely on my head.