Add support for DW_EH_PE_absptr, stub out support for DW_EH_PE_*rel values.
This adds a DWARF gnueh state context, which is used to encapsulate the
various base addresses against which relative pointer values will be
applied. The base addresses also support the new PL_VM_ADDRESS_INVALID
constant, which may be used to exclude base address types that are not
supported/available.

On Mac OS X, it doesn't appear that many of the relative address types
are used. Libunwind, for example, only supports pcrel and absptr.