Replace ambigious usage of pl_vm_* types with fixed-width types.
The pl_vm_address_t and pl_vm_size_t types are used to represent
addresses and lengths that are supported on the host architecture; eg,
they may be passed to Mach APIs for (re)mapping pages, etc.

This code, however, is host architecture agnostic, and is expected to
work even in the case where one parses a 64-bit target binary on a
32-bit host. This commit replaces all instances where data may be 32 or
64-bit with 64-bit fixed-width types.

All references to in-memory addresses continue to use pl_vm_address_t
offsets; even in the case where one is parsing a 64-bit binary, our
memory mappings will still be constrained by the limits of a 32-bit
host.

A future implementation could modify the mobject API to work on
files (rather than memory mapped data), in which case the
pl_vm_address_t and pl_vm_size_t types ought to be increased to 64-bit
width, with overflow validation performed in the memory mapping code
paths.

Such an implementation is relatively unlikely in PLCrashReporter itself,
as it is not intended to serve as a general purpose Mach-O / DWARF / CFE
decoding library; however, there's nothing that should prevent this code
being re-used for that purpose.