Clean up 32-bit/64-bit sign extension and overflow handling.
This fixes issues related to returning a 64-bit result type when
otherwise operating on 32-bit values, including 32-bit values that have
been casted from a signed type, and would require sign extension to be
properly represented in a 64-bit value.

Rather than require API clients to handle the sign extension themselves
while still using a generic 64-bit type, we expose _32 and _64-bit API
variants, thus making it explicit that the two are not interchangable,
and that API clients must select the appropriate variant based on the
data width with which they are operating. (These two functions both
pass through to the same template-based C++ implementation, avoiding
any actual code duplication).

In theory we could expose the C++ template itself, but I haven't yet
decided whether we should expose C++ APIs within the larger code base.

In addition to the above, this commit adds both 32-bit and 64-bit test
cases, which allow us to verify that the correct overflow/type width
handling occurs for both 32-bit and 64-bit targets, on both 32-bit and
64-bit hosts. On ARM, where no 64-bit thread state is supported by the
Mach APIs, we only test 32-bit operation. In the future, it may make
sense to divest ourselves of relying on Apple's thread state structures
directly, as this limits the target systems that the library can operate
on to whatever happens to be defined for the host system by Apple.