Lift dwarf_cfa_state out into the async-safe API group.
The aim here is to provide a re-usable set of unsupported, public APIs
that may be used to perform additional work at crash time by
plugins/callbacks registered with PLCrashReporter.

Since our API design requires exposing the dwarf_cfa_state class as part
of our CFA evaluation implementation, it needs to be lifted out such
that the headers can be published with the remainder of the framework.