Fix incorrect encoding of the pid/ppid.
Switches the encoding to the proto-defined uint32 type. As noted in the
commit comments, for correctness, this should be switched to int64 when
it's possible to deploy a breaking change to the protobuf serialization.

This commit also adds missing tests for the ProcessInfo fields (including
the pid fields), and enables AppInfo validation tests.

Issue: 66