Drop direct forwarding support, simplify the forwarding path.
This removes the broken direct forwarding code in favor of the more
typical 'proxy' forwarding path. The direct forwarding code was written to
explore ways of operating correctly even while running under a debugger,
however, the caveats were sufficiently intrusive that it seems
ineffective to pursue that approach.

To handle the case where the unit tests are running under a debugger
(which conflicts with our use of forwarding when testing the child
process handling, due to the debugger pausing our server thread), we
now reset the task/thread exception ports before executing our tests.