Add EXC_GUARD to the observed exception mask.
This was added in xnu 13.x+, and is triggered if an application
manipulates a file descriptor that has been guarded by the system.
See also: https://devforums.apple.com/message/713907#713907

We use runtime detection of the xnu version to avoid enabling this flag
on systems where it is not supported; this will return
KERN_INVALID_ARGUMENT, but the exact behavior of passing an
unsupported mask flag is undefined.

Issue: PLCR-468