Pass the correct thread and current thread state in the Mach exception path.
Previously, the current thread was always specified as the crashed
thread, and the crashed thread's state is provided as the current
thread's state.

This resulted in the following issues when using the Mach reporter:
- The mach exception thread was marked as the crashed thread.
- The mach exception thread was recorded with the crashed thread's stack
  frames.

The mach exception path now uses plcrash_async_thread_state_current() to
fetch and supply the mach exception thread state.

Issue: PLCR-494