Work around rdar://15042905 by mapping a segment other than __LINKEDIT.
On ARM64, The call to getsegmentdata() returns an invalid size value
that extends beyond the pages mapped for the segment; our tests crash
when memcmp() reads past the end of the mapped pages in the buffer
returned by getsegmentdata().

This seems to be caused by vmsize not being adjusted after the libPLCR
MH_OBJECT is linked with symbols stripped. If we disable symbol
stripping, the issue also disappears.

Whether this will impact anything beyond naive use of getsegmentdata()
is an open question, but since we can't ship with the private symbols
visible due to namespacing issues

The bug has been filed with Apple as rdar://15042905.

Issues: PLCR-495, PLCR-498