Fix the Mac OS X 10.8 and 10.9 builds.
This adds a PL_MIN_MACOSX_SDK project setting that is used to define the
minimum supported Mac OS X SDK; we'll continue to support building on
10.8, where various ARM64 defines are unavailable.

Once this is bumped to 10.9 -- or 10.8 SDK is updated to include
the ARM64 CPU types -- the replacement ARM #defines will trigger
compilation errors.

In addition, this commit also makes testing of objc_debug_isa_class_mask
conditional on __arm64__; the symbol is not exported on other
architectures.