Implement a clang-425.0.28-compatible work-around for -Wtautological-constant-out-of-range-compare.
In these particular cases, the point of the check is to document that
the issue has been (and in the future, should be) considered. The fact
that the cases are evaluated as tautological true is the point.

Specifying -Wno-tautological-constant-out-of-range-compare for the
entire project will break versions of clang that do not support the
warning flag (such as clang-425.0.28, shipped with Xcode 4). By using
the verbose pragma syntax, we can enable -Wno-unknown-pragmas for just
the code in question, which will allow clang to ignore subsequent the
'-Wtautological-constant-out-of-range-compare' flag if unknown.

This commit also adds a PLCR_PRAGMA_CLANG macro to enable/disable
clang-specific pragmas depending on the host compiler; the aim is to
keep the DWARF code modified here non-specific to clang.