Remove the simulator's 'run test' build phase.
In Xcode 5, the standard debug messages that fire when testing error
conditions are all treated as errors when they occur during the build
process.

We'll need to switch over to XCTest eventually, but in the meantime,
this replaces the complex instantiation of the unit tests within the
simulator, and instead matches the behavior to that of the device
tests, where we simply launch the test application directly.