Implement support for ARM64 DWARF, FRAME, and FRAMELESS encodings.
The unit tests validate the encoding as we understand it, but there may
be incompatibilities that remain between the actual in-use CFE encodings
and our implementation, especially as the ARM64 compact unwind
documentation is known to be inaccurate.

The unwinder appears to work against live CFE data, but additional integration
testing of the three encodings will be required prior to merging this
implementation to master.