Add PLCRASH_ARM64_FP to the list of non-volatile registers.
As per AAPCS64 Section 5.2.3, "A platform shall mandate the minimum
level of conformance with respect to the maintenance of frame records"

On iOS, Apple has defined the frame pointer behavior as follows: "The
frame pointer register (x29) must always address a valid frame record,
although some functions—such as leaf functions or tail calls—may elect
not to create an entry in this list. As a result, stack traces will
always be meaningful, even without debug information."

As such, we can treat the FP as a non-volatile register, callee saved.

In the case of leaf functions on ARM64, the compact unwind encoding does
not support restoration of the frame pointer, but -- as per Apple's
ARM64 Function Calling Conventions -- we may safely assume that the
frame pointer is left unmidified.