Fix overwrite kt value returned by vm_map().
This would have resulted in PLCRASH_EUSCCESS being returned in the case
that vm_map() returned failure, and the resulting mapped page range
would have been left as readable, mapped, zero-initialized data.

This would have avoided a crash (as the mapped range would in fact
be readable), but the cause of the invalid data would have been
misreported.