Set SKIP_INSTALL=yes for all embedded library targets.
When Xcode performs build+archive, it actually installs any targets for
which SKIP_INSTALL=no in their defined INSTALL_PATH:

* The static libraries are installed in DESTROOT/usr/local/lib
* The Mac OS X framework triggers an Xcode warning, as it has an
* INSTALL_PATH of @rpath
* plcrashutil is installed in /usr/local/bin
* etc.

This results in these libraries being incorrectly included in the output
of Build + Archive if PLCrashReporter is used as a nested project
dependency.

To resolve the issue, this commit sets SKIP_INSTALL=yes on all library
targets; all are meant to be used as embedded frameworks or static
libraries, and as such, they shouldn't ever be installed in a common
DESTROOT (and in the case of the Mac OS X frameworks, attempting to copy
them to "@rpath" is an error).