Implement the first proposed fix in PLCR-521; we'll be rolling this back.
This is untested and primarily commited for posterity. The risk here
ought to be obvious; by introducing a potential state in which we refuse
terminate a thread, we introduce the risk that a bug (such as a double
fault on the main thread) could lock up the process indefinitely.

Given that, this solution is too risky to be viable; we'll be revisiting
the question of safely generating temporary paths at crash time, such
that all writing paths may proceed concurrently.

Issue: PLCR-521