Expand the API documentation, and enable the ATTRIBUTE define on all compilers.
This further defines usage constraints and invariants, but does not yet
add the necessary documentation regarding write atomicity requirements
when updating the PLCrashImageAnnotation.

In addition, since the failure of the PLCRASH_IMAGE_ANNOTATION_ATTRIBUTE
would result in silent failure of PLCrashReporter at crash time (it
would simply not find the annotation), this commit removes the clang
check, and always defines the annotation -- if this breaks in another
compiler, it will at least do so visibly.

Issue: PLCR-527