Use the PLCrashProcessInfo API for fetching and validating process data from the crash report.
This allows us to use a common code path for fetching the process info,
rather than having a test-only code path that does not safely handle
UTF-8.

If we change the unit test's product name (as per the PLCR-532
reproduction case) to UTF-8 that triggers the multibyte sequence
truncation bug, this change allows the PLCrashLogWriterTests to complete
successfully.

Issue: PLCR-532