Work-around for rdar://16620815: iPhoneSimulator.platform ignores Xcode "Enable (C++|Objective-C) Exceptions" flags.
We rely on -fno-exceptions to avoid creating runtime link dependencies
on the C++ stdlib in our static library targets.

When building against the Simulator platform SDK, however, the
"Enable C++ Exceptions" project setting is ignored, breaking linking
of our static libraries in applications that do not not already
link libc++.

This is caused by broken compiler flag configurations in Xcode's
iPhoneSimulator.platform compiler specs; the compiler specs refer to a
legacy "GCC_ENABLE_EXCEPTIONS" flag that is no longer set by Xcode.

To work around this, we add "GCC_ENABLE_EXCEPTIONS" as a user-defined
build setting for the Simulator static library target, with a value
of NO. This hack should be removed once we stop supporting Xcode
releases with a broken Native Build System.xcspec; if the configuration
is not removed, however, it will likely operate as a no-op.

On the remote chance that the configuration does negatively impact the
build in a future Xcode release, it will only affect the simulator
targets.

Issue: PLCR-538