Fix AArch64 code alignment issues.
We missed a number of .align statements in the arm64 port. This happened
to work anyway, but in a few cases, misalignment is now triggering
linker warnings, alerting us to the bug.

This adds missing .align directives for all ARM64 code, and also fixes
an innocuous over-alignment (.align 4) for 32-bit ARM.

Issue: PLCR-542