Fix UNWIND_ARM64_MODE_* compatibility behavior for iOS targets.
These constants are defined as enums on iOS, preventing a stable #ifdef
check. As such, we always define these values on any Mac OS X target,
potentially overriding the existing enum identifiers.

To ensure this doesn't break inclusion of the header in which the enum
identifiers may be defined, we explicitly include the compact
unwind header.

Issue: PLCR-547