Adding a parser for DWARF EH LSDA records.
This currently is not async-safe due to the use of std::vector. Once a suitable async-safe data type is available, it should be switched over.

Additionally, the use of vector has forced the iOS demo app to link against libstdc++. This seems far from ideal and should be fixed.

Issue: PLCR-561