Use the new SpinLock class to support thread-safe usage of AsyncAllocator.
Note that this introduces an "async-safe" critical section; if the
process crashes within the allocator with a lock held, all other
allocation operations will deadlock if the same allocator is used.

To avoid this, it is recommended that allocators *not* be re-used within
a double-fault exception handler. We may also consider avoiding re-use
of allocators in the primary single fault exception handler.

Issue: PLCR-483