Integrate our kernel-level reference counting implementation.
This will serve as the basis for an async-safe shared_ptr/weak_ptr
reference counting implementation similar in design to the
C++11 APIs of the same name.

This code was previously written by myself for kernel-level use, and
the integration here is mostly unmodified from that code; the
implementation has not yet been attached to our async-safe allocator.

Issue: PLCR-570