Work-around a bug observed in iOS 9 beta's dyld
A company using PLCrashReporter noticed this behavior in the wild; it
appears that somehow -- possibly via broken dyld cache construction that
leaves the dyld command length running into zero-initialized padding --
some binaries contained LC_CMDs with zero length.

While this is absolutely invalid for Mach-O, we need to make progress
even in the face of broken dyld. To resolve this, we simply terminate
iteration when an invalid LC_CMD is hit; the reporting party has
confirmed that this resolves the issue.

If my guess regarding the nature of the dyld bug is correct (load command
section length extending into padding), then terminating iteration will
result in no impact on actual function.

However, and while there's no indication that this is the case, if the
terminated iteration prevents us from finding requisite load comments,
we will fail to parse the Mach-O file and the backtrace and symbols may
be left incomplete.

Issue: PLCR-579