Fix invalid return type of mach exception handler
The return type of this function is `kern_return_t`，however line 298 returns `false`  which means `KERN_SUCCESS` according to the definition `#define KERN_SUCCESS                    0` when writing report fails. It's supposed to be `KERN_FAILURE` rather than `KERN_SUCCESS` here. 

In some circumstances, the report folder in `Library/Caches` might be cleaned, so when an exception is thrown, the report would not be written successfully. However `mach_exception_callback` returns `KERN_SUCCESS` when writing the report. The thread would be resumed and exception would be thrown again and again.